add missing method for workspace client interface
