+++ src/plugins/workspace/public/components/workspace_form/types.ts
@@ -8,7 +8,7 @@ import type {
   ApplicationStart,
   SavedObjectsStart,
 } from '../../../../../core/public';
-import type { WorkspacePermissionMode } from '../../../common/constants';
+import type { WorkspacePermissionMode } from '../../../../../core/types';
 import type { WorkspaceOperationType, WorkspacePermissionItemType } from './constants';
 import { DataSourceConnection } from '../../../common/types';
 import { DataSourceManagementPluginSetup } from '../../../../../plugins/data_source_management/public';
