+++ src/plugins/workspace/server/routes/index.ts
@@ -7,7 +7,6 @@ import { schema } from '@osd/config-schema';
 import { IRouter, Logger, PrincipalType, ACL, DEFAULT_NAV_GROUPS } from '../../../../core/server';
 import { getUseCaseFeatureConfig } from '../../common/utils';
 import {
-  WorkspacePermissionMode,
   MAX_WORKSPACE_NAME_LENGTH,
   MAX_WORKSPACE_DESCRIPTION_LENGTH,
 } from '../../common/constants';
@@ -16,6 +15,7 @@ import { SavedObjectsPermissionControlContract } from '../permission_control/cli
 import { registerDuplicateRoute } from './duplicate';
 import { transferCurrentUserInPermissions, translatePermissionsToRole } from '../utils';
 import { validateWorkspaceColor } from '../../common/utils';
+import { WorkspacePermissionMode } from '../../../../core/types';
 
 export const WORKSPACES_API_BASE_URL = '/api/workspaces';
