+++ src/plugins/workspace/server/types.ts
@@ -3,8 +3,8 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
+import { WorkspaceFindOptions } from '../../../core/types';
 import {
-  Logger,
   OpenSearchDashboardsRequest,
   SavedObjectsFindResponse,
   CoreSetup,
@@ -17,17 +17,6 @@ import { PermissionModeId } from '../../../core/server';
 export interface WorkspaceAttributeWithPermission extends WorkspaceAttribute {
   permissions?: Permissions;
   permissionMode?: PermissionModeId;
-}
-import { WorkspacePermissionMode } from '../common/constants';
-
-export interface WorkspaceFindOptions {
-  page?: number;
-  perPage?: number;
-  search?: string;
-  searchFields?: string[];
-  sortField?: string;
-  sortOrder?: string;
-  permissionModes?: WorkspacePermissionMode[];
 }
 
 export interface IRequestDetail {
