+++ src/plugins/workspace/server/workspace_client.ts
@@ -17,7 +17,6 @@ import {
 import { updateWorkspaceState, getWorkspaceState } from '../../../core/server/utils';
 import {
   IWorkspaceClientImpl,
-  WorkspaceFindOptions,
   IResponse,
   IRequestDetail,
   WorkspaceAttributeWithPermission,
@@ -32,6 +31,7 @@ import {
   DATA_SOURCE_SAVED_OBJECT_TYPE,
   DATA_CONNECTION_SAVED_OBJECT_TYPE,
 } from '../../data_source/common';
+import { WorkspaceFindOptions } from '../../../core/types';
 
 const WORKSPACE_ID_SIZE = 6;
